meta-google: layer.conf: set BBFILE_PATTERN
Bitbake uses this regex to associate recipes and bbappends from BBFILES
(which contains recipes and bbappends from -all- layers listed in
bblayers.conf) to this layer.  The association is then used to map the
BBFILE_PRIORITY value for this layer onto the recipes provided by this
layer.

BBFILE_PRIORITY is used to determine which recipe to use when the same
recipe appears in multiple layers.

Without BBFILE_PATTERN set, recipes in this layer will get a default
priority and not the priority specified in BBFILE_PRIORITY.

This layer doesn't set a priority anyway, so functionally this patch is
a noop but it is an improvement from a correctness standpoint, and it
makes a warning go away

Removes warning:
WARNING: No bb files matched BBFILE_PATTERN_google-layer ''

Signed-off-by: Patrick Venture <venture@google.com>
Change-Id: I92ce6c0c8f05f75604d546fa168efe89532e96db