Add Ed as a meta-google maintainer
I really would've preferred pulling a sword from a stone, or been given a
2 factor encryption key by the lady of the lake, but as monty python
says "strange women lying in ponds distributing swords is no basis for a
system of government. Supreme executive power derives from a mandate
from the masses, not from some farcical aquatic ceremony."

As such, I submit this to the masses for approval.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I66cc689e5988b66356a97e9bc37b1bf63fe34aed